Introduce spring and java 21, refactored some code and removed unsued code
