+++ src/test/java/uk/gov/justice/digital/hmpps/vsip/annotation/ComponentWithWebDriver.java
@@ -17,5 +17,5 @@ import java.lang.annotation.Target;
 @Documented
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface LazyComponent {
+public @interface ComponentWithWebDriver {
 }
