--- src/test/java/uk/gov/justice/digital/hmpps/vsip/annotation/ElapsedTime.java
@@ -1,13 +1,0 @@
-package uk.gov.justice.digital.hmpps.vsip.annotation;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Documented
-@Target({ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ElapsedTime {
-}
