+++ src/test/java/uk/gov/justice/digital/hmpps/vsip/annotation/LazyConfiguration.java
@@ -1,6 +1,7 @@
 package uk.gov.justice.digital.hmpps.vsip.annotation;
 
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Lazy;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -8,6 +9,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+@Lazy
 @Configuration
 @Documented
 @Target({ElementType.TYPE})
