--- src/test/java/uk/gov/justice/digital/hmpps/vsip/annotation/SeleniumTest.java
@@ -1,18 +1,0 @@
-package uk.gov.justice.digital.hmpps.vsip.annotation;
-
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import uk.gov.justice.digital.hmpps.vsip.SpringSeleniumApplication;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.TYPE, ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-@ExtendWith(SpringExtension.class)
-@SpringBootTest(classes = SpringSeleniumApplication.class)
-public @interface SeleniumTest {
-}
