--- src/test/java/uk/gov/justice/digital/hmpps/vsip/annotation/TakeScreenshot.java
@@ -1,14 +1,0 @@
-package uk.gov.justice.digital.hmpps.vsip.annotation;
-
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Documented
-@Target({ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface TakeScreenshot {
-}
