+++ src/test/java/uk/gov/justice/digital/hmpps/vsip/config/SpringBeansConfigs.java
@@ -1,33 +1,21 @@
 package uk.gov.justice.digital.hmpps.vsip.config;
 
-import org.openqa.selenium.JavascriptExecutor;
-import org.openqa.selenium.WebDriver;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Scope;
 import uk.gov.justice.digital.hmpps.vsip.annotation.LazyAutowired;
 import uk.gov.justice.digital.hmpps.vsip.annotation.LazyConfiguration;
 import uk.gov.justice.digital.hmpps.vsip.util.Configuration;
 import uk.gov.justice.digital.hmpps.vsip.util.Environment;
-import uk.gov.justice.digital.hmpps.vsip.services.TestService;
+import uk.gov.justice.digital.hmpps.vsip.services.TestContextService;
 
 @LazyConfiguration
 public class SpringBeansConfigs {
 
     @LazyAutowired
-    private TestService testService;
-
+    private TestContextService testContextService;
 
     @Value("${environment:STAGING}")
     private Environment environment;
-
-    @Bean
-    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
-    public JavascriptExecutor javascriptExecutor(WebDriver driver) {
-        return (JavascriptExecutor) driver;
-    }
-
 
     @Bean
     public Configuration getConfiguration() {
