+++ src/test/java/uk/gov/justice/digital/hmpps/vsip/config/WebDriverConfig.java
@@ -1,6 +1,13 @@
 package uk.gov.justice.digital.hmpps.vsip.config;
 
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.edge.EdgeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.remote.LocalFileDetector;
+import org.openqa.selenium.remote.RemoteWebDriver;
+import org.openqa.selenium.safari.SafariDriver;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,38 +17,94 @@ import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.Scope;
-import uk.gov.justice.digital.hmpps.vsip.annotation.LazyAutowired;
 import uk.gov.justice.digital.hmpps.vsip.annotation.LazyConfiguration;
 import uk.gov.justice.digital.hmpps.vsip.annotation.WebdriverScopeBean;
-import uk.gov.justice.digital.hmpps.vsip.services.WebDriverService;
 
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.time.Duration;
+import java.util.concurrent.TimeUnit;
 
 @LazyConfiguration
 public class WebDriverConfig {
 
     @Value("${default.timeout:20}")
     private int timeout;
-    @LazyAutowired
-    private WebDriverService webDriverService;
+
+
+    private static final Logger LOG = LoggerFactory.getLogger(WebDriverConfig.class);
 
     /**
      * Method as static to allow the bean to be created without first creating an instance of your @Configuration
      **/
     @Bean
     public static BeanFactoryPostProcessor beanFactoryPostProcessor() {
-        return new WebdriverScopePostProcessor();
+        return new WebDriverScopePostProcessor();
     }
 
     @WebdriverScopeBean
     @Primary
     public WebDriver createDriverBean() {
-        return webDriverService.getWebDriver();
+        return getWebDriver();
     }
 
+    /**
+     * The prototype scope type will return a different bean everytime it is autowired
+     */
     @Bean
     @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
     public WebDriverWait webdriverWait(WebDriver driver) {
         return new WebDriverWait(driver, Duration.ofMillis(this.timeout));
+    }
+
+
+    @Value("${browser:chrome}")
+    private String targetBrowser;
+
+    private WebDriver getWebDriver() {
+
+        final WebDriver webDriver;
+        switch (targetBrowser.toLowerCase()) {
+            case "chrome":
+                webDriver = new ChromeDriver();
+                break;
+            case "firefox":
+                webDriver = new FirefoxDriver();
+                break;
+            case "safari":
+                webDriver = new SafariDriver();
+                break;
+            case "edge":
+                webDriver = new EdgeDriver();
+                break;
+            case "remote-chrome":
+            case "remote-firefox":
+                webDriver = createRemoteDriver();
+                break;
+            default:
+                throw new IllegalArgumentException("Target browser " + targetBrowser + " not recognized");
+        }
+
+        webDriver.manage().window().maximize();
+        webDriver.manage().timeouts().pageLoadTimeout(60, TimeUnit.SECONDS);
+        webDriver.manage().timeouts().setScriptTimeout(60, TimeUnit.SECONDS);
+        webDriver.manage().timeouts().implicitlyWait(20, TimeUnit.SECONDS);
+        webDriver.getWindowHandle();
+
+        LOG.debug("Enter createDriverBean , Created Webdriver :" + webDriver.getClass().getSimpleName() + "/" + targetBrowser);
+
+        return webDriver;
+    }
+
+    private WebDriver createRemoteDriver() {
+        try {
+            ChromeOptions options = new ChromeOptions();
+            RemoteWebDriver driver = new RemoteWebDriver(new URL("http://localhost:4444/wd/hub"), options);
+            driver.setFileDetector(new LocalFileDetector());
+            return driver;
+        } catch (MalformedURLException e) {
+            throw new RuntimeException("Failed to create RemoteWebDriver instance", e);
+        }
+
     }
 }
