+++ src/test/java/uk/gov/justice/digital/hmpps/vsip/config/WebDriverScope.java
@@ -7,7 +7,10 @@ import org.springframework.context.support.SimpleThreadScope;
 
 import java.util.Objects;
 
-public class WebdriverScope extends SimpleThreadScope {
+public class WebDriverScope extends SimpleThreadScope {
+
+    public static final String WEB_DRIVER_SCOPE = "webdriverscope";
+
     @Override
     public Object get(String name, ObjectFactory<?> objectFactory) {
         Object object = super.get(name, objectFactory);
@@ -15,7 +18,8 @@ public class WebdriverScope extends SimpleThreadScope {
             SessionId sessionId = remoteWebDriver.getSessionId();
             if (Objects.isNull(sessionId)) {
                 super.remove(name);
-                object = super.get(name, objectFactory);
+                // I assume this is when the session expires, recreate a new one
+                return super.get(name, objectFactory);
             }
         }
         return object;
