+++ src/test/java/uk/gov/justice/digital/hmpps/vsip/config/WebDriverScopePostProcessor.java
@@ -4,9 +4,12 @@ import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
 
-public class WebdriverScopePostProcessor implements BeanFactoryPostProcessor {
+import static uk.gov.justice.digital.hmpps.vsip.config.WebDriverScope.WEB_DRIVER_SCOPE;
+
+public class WebDriverScopePostProcessor implements BeanFactoryPostProcessor {
+
     @Override
     public void postProcessBeanFactory(ConfigurableListableBeanFactory beanFactory) throws BeansException {
-        beanFactory.registerScope("webdriverscope", new WebdriverScope());
+        beanFactory.registerScope(WEB_DRIVER_SCOPE, new WebDriverScope());
     }
 }
