+++ src/test/java/uk/gov/justice/digital/hmpps/vsip/cucumber/CucumberHooksConfig.java
@@ -3,6 +3,7 @@ package uk.gov.justice.digital.hmpps.vsip.cucumber;
 import io.cucumber.java.After;
 import io.cucumber.java.AfterStep;
 import io.cucumber.java.Before;
+import io.cucumber.java.BeforeStep;
 import io.cucumber.java.Scenario;
 import org.openqa.selenium.WebDriver;
 import org.slf4j.Logger;
@@ -10,7 +11,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ApplicationContext;
 import uk.gov.justice.digital.hmpps.vsip.annotation.LazyAutowired;
-import uk.gov.justice.digital.hmpps.vsip.services.TestService;
+import uk.gov.justice.digital.hmpps.vsip.services.TestContextService;
 
 /**
  * This hooks class must be in the same package as the CucumberSpringContextConfig
@@ -23,7 +24,7 @@ public class CucumberHooksConfig {
     private String targetBrowser;
 
     @LazyAutowired
-    private TestService testService;
+    private TestContextService testContextService;
 
     @LazyAutowired
     private ApplicationContext applicationContext;
@@ -31,17 +32,26 @@ public class CucumberHooksConfig {
     @LazyAutowired
     private WebDriver webDriver;
 
+    /**
+     * Called for every Scenario
+     */
     @Before
     public void initialise(){
         LOG.debug("Entered initialise deleteAllCookies");
         webDriver.manage().deleteAllCookies();
+        testContextService.clearTestContext();
+    }
+
+    @BeforeStep
+    public void beforeStep(Scenario scenario) {
+
     }
 
     @AfterStep
     public void afterStep(Scenario scenario) {
         if (scenario.isFailed()) {
             LOG.debug("Entered afterStep" + scenario.getName() + " Failed therefore take screen shot");
-            scenario.attach(this.testService.getScreenshot(), "image/png", scenario.getName());
+            scenario.attach(this.testContextService.getScreenshot(), "image/png", scenario.getName());
         }
     }
