+++ src/test/java/uk/gov/justice/digital/hmpps/vsip/cucumber/steps/ConfirmationSteps.java
@@ -4,7 +4,7 @@ import io.cucumber.java.en.And;
 import uk.gov.justice.digital.hmpps.vsip.annotation.LazyAutowired;
 import uk.gov.justice.digital.hmpps.vsip.pages.ConfirmationPage;
 import uk.gov.justice.digital.hmpps.vsip.util.Context;
-import uk.gov.justice.digital.hmpps.vsip.services.TestService;
+import uk.gov.justice.digital.hmpps.vsip.services.TestContextService;
 
 public class ConfirmationSteps {
 
@@ -12,7 +12,7 @@ public class ConfirmationSteps {
     private ConfirmationPage confirmationPage;
 
     @LazyAutowired
-    private TestService testService;
+    private TestContextService testContextService;
 
     @And("I see {string} message displayed")
     public void iSeeMessageDisplayed(String confirmationMsg) {
@@ -22,7 +22,7 @@ public class ConfirmationSteps {
     @And("I see a booking reference")
     public void iSeeABookingReferenceDisplayed() {
         String bookingReference = confirmationPage.getBookingReference();
-        TestService.setToContext(Context.BOOKING_REFERENCE, bookingReference);
+        testContextService.setToContext(Context.BOOKING_REFERENCE, bookingReference);
     }
