+++ src/test/java/uk/gov/justice/digital/hmpps/vsip/pages/AdditionalSupportPage.java
@@ -1,8 +1,8 @@
 package uk.gov.justice.digital.hmpps.vsip.pages;
 
-import uk.gov.justice.digital.hmpps.vsip.annotation.LazyComponent;
+import uk.gov.justice.digital.hmpps.vsip.annotation.ComponentWithWebDriver;
 
-@LazyComponent
+@ComponentWithWebDriver
 public class AdditionalSupportPage extends BasePage {
 
     public void selectNoForAddAssistance() {
